drivers/requirements.txt
labjack-ljm
pymodbus
psycopg2
python-dotenv